+++ src/client-side-metrics/gcp-metrics-handler.ts
@@ -22,7 +22,6 @@ import * as Resources from '@opentelemetry/resources';
 import * as ResourceUtil from '@google-cloud/opentelemetry-resource-util';
 import {PushMetricExporter, View} from '@opentelemetry/sdk-metrics';
 import {ClientOptions} from 'google-gax';
-import {generateClientUuid} from './generate-client-uuid';
 const {
   Aggregation,
   ExplicitBucketHistogramAggregation,
@@ -30,6 +29,31 @@ const {
   Histogram,
   PeriodicExportingMetricReader,
 } = require('@opentelemetry/sdk-metrics');
+import * as os from 'os';
+import * as crypto from 'crypto';
+
+/**
+ * Generates a unique client identifier string.
+ *
+ * This function creates a client identifier that incorporates the hostname,
+ * process ID, and a UUID to ensure uniqueness across different client instances
+ * and processes. The identifier follows the pattern:
+ *
+ * `node-<uuid>-<pid><hostname>`
+ *
+ * where:
+ * - `<uuid>` is a randomly generated UUID (version 4).
+ * - `<pid>` is the process ID of the current Node.js process.
+ * - `<hostname>` is the hostname of the machine.
+ *
+ * @returns {string} A unique client identifier string.
+ */
+function generateClientUuid() {
+  const hostname = os.hostname() || 'localhost';
+  const currentPid = process.pid || '';
+  const uuid4 = crypto.randomUUID();
+  return `node-${uuid4}-${currentPid}${hostname}`;
+}
 
 /**
  * A collection of OpenTelemetry metric instruments used to record
@@ -73,7 +97,7 @@ function createInstruments(exporter: PushMetricExporter): MetricsInstruments {
       new View({
         instrumentName: name,
         name,
-        aggregation: name.endsWith('latencies')
+        aggregation: !name.endsWith('latencies')
           ? Aggregation.Sum()
           : new ExplicitBucketHistogramAggregation(latencyBuckets),
       }),