+++ system-test/client-side-metrics.ts
@@ -23,6 +23,7 @@ import * as assert from 'assert';
 import {GCPMetricsHandler} from '../src/client-side-metrics/gcp-metrics-handler';
 import * as proxyquire from 'proxyquire';
 import {Bigtable} from '../src';
+import {Row} from '../src/row';
 import {setupBigtable} from './client-side-metrics-setup-table';
 import {TestMetricsHandler} from '../test-common/test-metrics-handler';
 import {
@@ -30,8 +31,8 @@ import {
   OnOperationCompleteData,
 } from '../src/client-side-metrics/metrics-handler';
 import {ClientOptions} from 'google-gax';
-import {PassThrough} from 'stream';
-import {generateChunksFromRequest} from '../test-common/utils/readRowsImpl';
+import {ClientSideMetricsConfigManager} from '../src/client-side-metrics/metrics-config-manager';
+import {MetricServiceClient} from '@google-cloud/monitoring';
 
 const SECOND_PROJECT_ID = 'cfdb-sdk-node-tests';
 
@@ -56,29 +57,18 @@ class FakeHRTime {
 function getFakeBigtable(
   projectId: string,
   metricsHandlerClass: typeof GCPMetricsHandler | typeof TestMetricsHandler,
-  hrtime: FakeHRTime,
+  apiEndpoint?: string,
 ) {
-  const metricHandler = new metricsHandlerClass(
-    {} as unknown as ClientOptions & {value: string},
-  );
-  const FakeOperationsMetricsCollector = proxyquire(
-    '../src/client-side-metrics/operation-metrics-collector.js',
-    {
-      'node:process': {
-        hrtime,
-      },
-    },
-  ).OperationMetricsCollector;
-  const FakeClientSideMetricsConfigManager = proxyquire(
-    '../src/client-side-metrics/metrics-config-manager.js',
-    {
-      './operation-metrics-collector.js': {
-        OperationMetricsCollector: FakeOperationsMetricsCollector,
-      },
-    },
-  ).ClientSideMetricsConfigManager;
-  const newClient = new Bigtable({projectId});
-  newClient._metricsConfigManager = new FakeClientSideMetricsConfigManager([
+  // Normally the options passed into the client are passed into the metrics
+  // handler so when we mock out the metrics handler, it really should have
+  // the same options that are passed into the client.
+  const options = {
+    projectId,
+    apiEndpoint,
+  };
+  const metricHandler = new metricsHandlerClass(options);
+  const newClient = new Bigtable(options);
+  newClient._metricsConfigManager = new ClientSideMetricsConfigManager([
     metricHandler,
   ]);
   return newClient;
@@ -92,7 +82,179 @@ function getHandlerFromExporter(Exporter: typeof CloudMonitoringExporter) {
   }).GCPMetricsHandler;
 }
 
-describe.only('Bigtable/ClientSideMetrics', () => {
+function readRowsAssertionCheck(
+  projectId: string,
+  requestsHandled: (OnOperationCompleteData | OnAttemptCompleteData)[] = [],
+  method: string,
+  streaming: string,
+) {
+  assert.strictEqual(requestsHandled.length, 4);
+  const firstRequest = requestsHandled[0] as any;
+  // We would expect these parameters to be different every time so delete
+  // them from the comparison after checking they exist.
+  assert(firstRequest.attemptLatency);
+  assert(firstRequest.serverLatency);
+  delete firstRequest.attemptLatency;
+  delete firstRequest.serverLatency;
+  delete firstRequest.metricsCollectorData.appProfileId;
+  assert.deepStrictEqual(firstRequest, {
+    connectivityErrorCount: 0,
+    streaming,
+    status: '0',
+    client_name: 'nodejs-bigtable',
+    metricsCollectorData: {
+      instanceId: 'emulator-test-instance',
+      table: 'my-table',
+      cluster: 'fake-cluster3',
+      zone: 'us-west1-c',
+      method,
+    },
+    projectId,
+  });
+  const secondRequest = requestsHandled[1] as any;
+  // We would expect these parameters to be different every time so delete
+  // them from the comparison after checking they exist.
+  assert(secondRequest.operationLatency);
+  assert(secondRequest.firstResponseLatency);
+  assert(secondRequest.applicationLatencies);
+  delete secondRequest.operationLatency;
+  delete secondRequest.firstResponseLatency;
+  delete secondRequest.applicationLatencies;
+  delete secondRequest.metricsCollectorData.appProfileId;
+  assert.deepStrictEqual(secondRequest, {
+    status: '0',
+    streaming,
+    client_name: 'nodejs-bigtable',
+    metricsCollectorData: {
+      instanceId: 'emulator-test-instance',
+      cluster: 'fake-cluster3',
+      zone: 'us-west1-c',
+      method,
+      table: 'my-table',
+    },
+    projectId,
+    retryCount: 0,
+  });
+  // We would expect these parameters to be different every time so delete
+  // them from the comparison after checking they exist.
+  const thirdRequest = requestsHandled[2] as any;
+  assert(thirdRequest.attemptLatency);
+  assert(thirdRequest.serverLatency);
+  delete thirdRequest.attemptLatency;
+  delete thirdRequest.serverLatency;
+  delete thirdRequest.metricsCollectorData.appProfileId;
+  assert.deepStrictEqual(thirdRequest, {
+    connectivityErrorCount: 0,
+    streaming,
+    status: '0',
+    client_name: 'nodejs-bigtable',
+    metricsCollectorData: {
+      instanceId: 'emulator-test-instance',
+      table: 'my-table2',
+      cluster: 'fake-cluster3',
+      zone: 'us-west1-c',
+      method,
+    },
+    projectId,
+  });
+  const fourthRequest = requestsHandled[3] as any;
+  // We would expect these parameters to be different every time so delete
+  // them from the comparison after checking they exist.
+  assert(fourthRequest.operationLatency);
+  assert(fourthRequest.firstResponseLatency);
+  assert(fourthRequest.applicationLatencies);
+  delete fourthRequest.operationLatency;
+  delete fourthRequest.firstResponseLatency;
+  delete fourthRequest.applicationLatencies;
+  delete fourthRequest.metricsCollectorData.appProfileId;
+  assert.deepStrictEqual(fourthRequest, {
+    status: '0',
+    streaming,
+    client_name: 'nodejs-bigtable',
+    metricsCollectorData: {
+      instanceId: 'emulator-test-instance',
+      cluster: 'fake-cluster3',
+      zone: 'us-west1-c',
+      method,
+      table: 'my-table2',
+    },
+    projectId,
+    retryCount: 0,
+  });
+}
+
+function checkMultiRowCall(
+  projectId: string,
+  requestsHandled: (OnOperationCompleteData | OnAttemptCompleteData)[] = [],
+) {
+  readRowsAssertionCheck(
+    projectId,
+    requestsHandled,
+    'Bigtable.ReadRows',
+    'true',
+  );
+}
+
+function checkSingleRowCall(
+  projectId: string,
+  requestsHandled: (OnOperationCompleteData | OnAttemptCompleteData)[] = [],
+) {
+  readRowsAssertionCheck(
+    projectId,
+    requestsHandled,
+    'Bigtable.ReadRow',
+    'false',
+  );
+}
+
+/**
+ * Checks if metrics have been published to Google Cloud Monitoring.
+ *
+ * This asynchronous function queries Google Cloud Monitoring to verify
+ * that the expected metrics from the Bigtable client library have been
+ * successfully published. It constructs a `MetricServiceClient` to
+ * interact with the Cloud Monitoring API and retrieves time series data
+ * for a predefined set of metrics. The test passes if time series data
+ * is found for each of the specified metrics within a defined time
+ * interval.
+ *
+ * @param {string} projectId The Google Cloud project ID where metrics are
+ *   expected to be published.
+ * @throws {Error} If no time series data is found for any of the specified
+ *   metrics, indicating that the metrics were not successfully published to
+ *   Cloud Monitoring.
+ */
+async function checkForPublishedMetrics(projectId: string) {
+  const monitoringClient = new MetricServiceClient({projectId}); // Correct instantiation
+  const now = Math.floor(Date.now() / 1000);
+  const filters = [
+    'metric.type="bigtable.googleapis.com/client/attempt_latencies"',
+    'metric.type="bigtable.googleapis.com/client/operation_latencies"',
+    'metric.type="bigtable.googleapis.com/client/retry_count"',
+    'metric.type="bigtable.googleapis.com/client/server_latencies"',
+    'metric.type="bigtable.googleapis.com/client/first_response_latencies"',
+  ];
+  for (let i = 0; i < filters.length; i++) {
+    const filter = filters[i];
+    const [series] = await monitoringClient.listTimeSeries({
+      name: `projects/${projectId}`,
+      interval: {
+        endTime: {
+          seconds: now,
+          nanos: 0,
+        },
+        startTime: {
+          seconds: now - 1000 * 60 * 60 * 24,
+          nanos: 0,
+        },
+      },
+      filter,
+    });
+    assert(series.length > 0);
+  }
+}
+
+describe('Bigtable/ClientSideMetrics', () => {
   const instanceId1 = 'emulator-test-instance';
   const instanceId2 = 'emulator-test-instance2';
   const tableId1 = 'my-table';
@@ -142,7 +304,11 @@ describe.only('Bigtable/ClientSideMetrics', () => {
   describe('Bigtable/ClientSideMetricsToGCM', () => {
     // This test suite ensures that for each test all the export calls are
     // successful even when multiple instances and tables are created.
-    async function mockBigtable(projectId: string, done: mocha.Done) {
+    async function mockBigtable(
+      projectId: string,
+      done: mocha.Done,
+      apiEndpoint?: string,
+    ) {
       /*
       The exporter is called every x seconds, but we only want to test the value
       it receives once. Since done cannot be called multiple times in mocha,
@@ -183,7 +349,14 @@ describe.only('Bigtable/ClientSideMetrics', () => {
                   // result from calling export was successful.
                   assert.strictEqual(result.code, 0);
                   resultCallback({code: 0});
-                  done();
+                  void checkForPublishedMetrics(projectId)
+                    .then(() => {
+                      done();
+                    })
+                    .catch(err => {
+                      done(new Error('Metrics have not been published'));
+                      done(err);
+                    });
                 } catch (error) {
                   // The code here isn't 0 so we report the original error to the mocha test runner.
                   done(result);
@@ -202,7 +375,7 @@ describe.only('Bigtable/ClientSideMetrics', () => {
       return getFakeBigtable(
         projectId,
         getHandlerFromExporter(TestExporter),
-        new FakeHRTime(),
+        apiEndpoint,
       );
     }
 
@@ -220,6 +393,32 @@ describe.only('Bigtable/ClientSideMetrics', () => {
             await table.getRows();
             const table2 = instance.table(tableId2);
             await table2.getRows();
+          }
+        } catch (e) {
+          done(new Error('An error occurred while running the script'));
+          done(e);
+        }
+      })().catch(err => {
+        throw err;
+      });
+    });
+    it('should send the metrics to Google Cloud Monitoring for a custom endpoint', done => {
+      (async () => {
+        try {
+          const bigtable = await mockBigtable(
+            defaultProjectId,
+            done,
+            'bogus-endpoint',
+          );
+          const instance = bigtable.instance(instanceId1);
+          const table = instance.table(tableId1);
+          try {
+            // This call will fail because we are trying to hit a bogus endpoint.
+            // The idea here is that we just want to record at least one metric
+            // so that the exporter gets executed.
+            await table.getRows();
+          } catch (e: unknown) {
+            // Try blocks just need a catch/finally block.
           }
         } catch (e) {
           done(new Error('An error occurred while running the script'));
@@ -274,7 +473,7 @@ describe.only('Bigtable/ClientSideMetrics', () => {
           resultCallback: (result: ExportResult) => void,
         ): Promise<void> {
           try {
-            await super.export(metrics, (result: ExportResult) => {
+            await super.export(metrics, async (result: ExportResult) => {
               try {
                 // The code is expected to be 0 because the
                 // result from calling export was successful.
@@ -306,11 +505,7 @@ describe.only('Bigtable/ClientSideMetrics', () => {
       ensure the export was successful and pass the test with code 0 if it is
       successful.
        */
-      return getFakeBigtable(
-        projectId,
-        getHandlerFromExporter(TestExporter),
-        new FakeHRTime(),
-      );
+      return getFakeBigtable(projectId, getHandlerFromExporter(TestExporter));
     }
 
     it('should send the metrics to Google Cloud Monitoring for a ReadRows call', done => {
@@ -625,11 +820,7 @@ describe.only('Bigtable/ClientSideMetrics', () => {
         }
       }
 
-      const bigtable = getFakeBigtable(
-        projectId,
-        TestGCPMetricsHandler,
-        hrtime,
-      );
+      const bigtable = getFakeBigtable(projectId, TestGCPMetricsHandler);
       await setupBigtable(bigtable, columnFamilyId, instanceId1, [
         tableId1,
         tableId2,
@@ -641,10 +832,9 @@ describe.only('Bigtable/ClientSideMetrics', () => {
       (async () => {
         const projectId = defaultProjectId;
         const bigtable = await mockBigtable(
-          projectId,
+          defaultProjectId,
           done,
-          standardAssertionChecks,
-          new FakeHRTime(),
+          checkMultiRowCall,
         );
         const instance = bigtable.instance(instanceId1);
         const table = instance.table(tableId1);
@@ -657,141 +847,36 @@ describe.only('Bigtable/ClientSideMetrics', () => {
     });
     it('should pass the projectId to the metrics handler properly', done => {
       (async () => {
+        const bigtable = await mockBigtable(
+          defaultProjectId,
+          done,
+          checkMultiRowCall,
+        );
+        const instance = bigtable.instance(instanceId1);
+        const table = instance.table(tableId1);
+        await table.getRows();
+        const table2 = instance.table(tableId2);
+        await table2.getRows();
+      })().catch(err => {
+        throw err;
+      });
+    });
+    it('should send the metrics to the metrics handler for a single row read', done => {
+      (async () => {
         try {
           const projectId = SECOND_PROJECT_ID;
           const bigtable = await mockBigtable(
             projectId,
             done,
-            standardAssertionChecks,
-            new FakeHRTime(),
+            checkSingleRowCall,
           );
           const instance = bigtable.instance(instanceId1);
           const table = instance.table(tableId1);
-          await table.getRows();
+          const row = new Row(table, 'rowId');
+          await row.get();
           const table2 = instance.table(tableId2);
-          await table2.getRows();
-        } catch (e) {
-          done(e);
-        }
-      })().catch(err => {
-        throw err;
-      });
-    });
-    it('should record the right metrics when handling rows through readrows stream', done => {
-      (async () => {
-        try {
-          const hrtime = new FakeHRTime();
-          const bigtable = await mockBigtable(
-            SECOND_PROJECT_ID,
-            done,
-            applicationLatenciesChecks,
-            hrtime,
-          );
-          const instance = bigtable.instance(instanceId1);
-          const table = instance.table(tableId1);
-          // Mock stream behaviour:
-          // @ts-ignore
-          table.bigtable.request = () => {
-            const chunks = generateChunksFromRequest(
-              {},
-              {
-                chunkSize: 1,
-                valueSize: 1,
-                errorAfterChunkNo: 2,
-                keyFrom: 0,
-                keyTo: 3,
-                chunksPerResponse: 1,
-                debugLog: () => {},
-              },
-            );
-            const data = {
-              lastRowKey: chunks[2].rowKey,
-              chunks,
-            };
-            const stream = new PassThrough({
-              objectMode: true,
-            });
-
-            setImmediate(() => {
-              stream.emit('data', data);
-              stream.emit('end');
-            });
-
-            return stream;
-          };
-          const stream = table.createReadStream();
-          stream.on('data', () => {
-            // Simulate an application that takes 5 seconds between row reads.
-            hrtime.bigint();
-            hrtime.bigint();
-            hrtime.bigint();
-            hrtime.bigint();
-            hrtime.bigint();
-          });
-          stream.on('end', async () => {
-            const table2 = instance.table(tableId2);
-            await table2.getRows();
-          });
-        } catch (e) {
-          done(e);
-        }
-      })().catch(err => {
-        throw err;
-      });
-    });
-    it('should record the right metrics when iterating through readrows stream', done => {
-      (async () => {
-        try {
-          const hrtime = new FakeHRTime();
-          const bigtable = await mockBigtable(
-            SECOND_PROJECT_ID,
-            done,
-            applicationLatenciesChecks,
-            hrtime,
-          );
-          const instance = bigtable.instance(instanceId1);
-          const table = instance.table(tableId1);
-          // Mock stream behaviour:
-          // @ts-ignore
-          table.bigtable.request = () => {
-            const chunks = generateChunksFromRequest(
-              {},
-              {
-                chunkSize: 1,
-                valueSize: 1,
-                errorAfterChunkNo: 2,
-                keyFrom: 0,
-                keyTo: 3,
-                chunksPerResponse: 1,
-                debugLog: () => {},
-              },
-            );
-            const data = {
-              lastRowKey: chunks[2].rowKey,
-              chunks,
-            };
-            const stream = new PassThrough({
-              objectMode: true,
-            });
-
-            setImmediate(() => {
-              stream.emit('data', data);
-              stream.emit('end');
-            });
-
-            return stream;
-          };
-          const stream = table.createReadStream();
-          for await (const row of stream) {
-            // Simulate an application that takes 5 seconds between row reads.
-            hrtime.bigint();
-            hrtime.bigint();
-            hrtime.bigint();
-            hrtime.bigint();
-            hrtime.bigint();
-          }
-          const table2 = instance.table(tableId2);
-          await table2.getRows();
+          const row2 = new Row(table2, 'rowId');
+          await row2.get();
         } catch (e) {
           done(e);
         }