+++ test/row.ts
@@ -19,9 +19,19 @@ import * as proxyquire from 'proxyquire';
 import * as sinon from 'sinon';
 import {Mutation} from '../src/mutation.js';
 import * as rw from '../src/row';
-import {Table, Entry} from '../src/table.js';
+import {
+  Table,
+  Entry,
+  GetRowsOptions,
+  GetRowsCallback,
+  GetRowsResponse,
+} from '../src/table.js';
 import {Chunk} from '../src/chunktransformer.js';
-import {CallOptions} from 'google-gax';
+import {CallOptions, ServiceError} from 'google-gax';
+import {ClientSideMetricsConfigManager} from '../src/client-side-metrics/metrics-config-manager';
+import {Bigtable} from '../src/';
+import {getRowsInternal} from '../src/utils/getRowsInternal';
+import {TabularApiSurface} from '../src/tabular-api-surface';
 
 const sandbox = sinon.createSandbox();
 
@@ -78,19 +88,36 @@ describe('Bigtable/Row', () => {
   let RowError: typeof rw.RowError;
   let row: rw.Row;
 
-  before(() => {
+  function getFakeRow(
+    getRowsInternal: (
+      table: TabularApiSurface,
+      singleRow: boolean,
+      optionsOrCallback?: GetRowsOptions | GetRowsCallback,
+      cb?: GetRowsCallback,
+    ) => void | Promise<GetRowsResponse>,
+  ) {
     const Fake = proxyquire('../src/row.js', {
       '@google-cloud/promisify': fakePromisify,
       './mutation.js': {Mutation: FakeMutation},
       './filter.js': {Filter: FakeFilter},
       './row-data-utils.js': {RowDataUtils: FakeRowDataUtil},
+      './utils/getRowsInternal': {
+        getRowsInternal,
+      },
     });
-    Row = Fake.Row;
     RowError = Fake.RowError;
+    return Fake;
+  }
+
+  before(() => {
+    const Fake = getFakeRow(() => {});
+    Row = Fake.Row;
   });
 
   beforeEach(() => {
     row = new Row(TABLE, ROW_ID);
+    row.table.bigtable._metricsConfigManager =
+      new ClientSideMetricsConfigManager([]);
   });
 
   afterEach(() => {
@@ -997,15 +1024,48 @@ describe('Bigtable/Row', () => {
   });
 
   describe('get', () => {
+    function getRowInstance(
+      fn: (reqOpts: any) => void | Promise<GetRowsResponse>,
+    ) {
+      const getRowsInternal = (
+        table: TabularApiSurface,
+        singleRow: boolean,
+        optionsOrCallback?: GetRowsOptions | GetRowsCallback,
+        cb?: GetRowsCallback,
+      ) => {
+        return fn(optionsOrCallback);
+      };
+      const Fake = getFakeRow(getRowsInternal);
+      Row = Fake.Row;
+      row = new Row(TABLE, ROW_ID);
+      return row;
+    }
+
+    function getRowInstanceForErrResp(err: ServiceError | null, resp?: any[]) {
+      const getRowsInternal = (
+        table: TabularApiSurface,
+        singleRow: boolean,
+        optionsOrCallback?: GetRowsOptions | GetRowsCallback,
+        cb?: GetRowsCallback,
+      ) => {
+        if (cb) {
+          cb(err, resp);
+        }
+      };
+      const Fake = getFakeRow(getRowsInternal);
+      Row = Fake.Row;
+      row = new Row(TABLE, ROW_ID);
+      return row;
+    }
     it('should provide the proper request options', done => {
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (row.table.getRows as Function) = (reqOpts: any) => {
+      const fn = (reqOpts: any) => {
         assert.strictEqual(reqOpts.keys[0], ROW_ID);
         assert.strictEqual(reqOpts.filter, undefined);
         assert.strictEqual(FakeMutation.parseColumnName.callCount, 0);
         done();
       };
-
+      const row = getRowInstance(fn);
       row.get(assert.ifError);
     });
 
@@ -1022,12 +1082,13 @@ describe('Bigtable/Row', () => {
       ];
 
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (row.table.getRows as Function) = (reqOpts: any) => {
+      const fn = (reqOpts: any) => {
         assert.deepStrictEqual(reqOpts.filter, expectedFilter);
         assert.strictEqual(FakeMutation.parseColumnName.callCount, 1);
         assert(FakeMutation.parseColumnName.calledWith(keys[0]));
         done();
       };
+      const row = getRowInstance(fn);
       row.get(keys, assert.ifError);
     });
 
@@ -1058,7 +1119,7 @@ describe('Bigtable/Row', () => {
       ];
 
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (row.table.getRows as Function) = (reqOpts: any) => {
+      const fn = (reqOpts: any) => {
         assert.deepStrictEqual(reqOpts.filter, expectedFilter);
 
         const spy = FakeMutation.parseColumnName;
@@ -1068,6 +1129,7 @@ describe('Bigtable/Row', () => {
         assert.strictEqual(spy.getCall(1).args[0], keys[1]);
         done();
       };
+      const row = getRowInstance(fn);
 
       row.get(keys, assert.ifError);
     });
@@ -1082,12 +1144,13 @@ describe('Bigtable/Row', () => {
       ];
 
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (row.table.getRows as Function) = (reqOpts: any) => {
+      const fn = (reqOpts: any) => {
         assert.deepStrictEqual(reqOpts.filter, expectedFilter);
         assert.strictEqual(FakeMutation.parseColumnName.callCount, 1);
         assert(FakeMutation.parseColumnName.calledWith(keys[0]));
         done();
       };
+      const row = getRowInstance(fn);
 
       row.get(keys, assert.ifError);
     });
@@ -1121,13 +1184,14 @@ describe('Bigtable/Row', () => {
       ];
 
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (row.table.getRows as Function) = (reqOpts: any) => {
+      const fn = (reqOpts: any) => {
         assert.deepStrictEqual(reqOpts.filter, expectedFilter);
         assert.strictEqual(FakeMutation.parseColumnName.callCount, 1);
         assert(FakeMutation.parseColumnName.calledWith(keys[0]));
         assert.strictEqual(reqOpts.decode, options.decode);
         done();
       };
+      const row = getRowInstance(fn);
 
       row.get(keys, options, assert.ifError);
     });
@@ -1175,13 +1239,14 @@ describe('Bigtable/Row', () => {
       ];
 
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (row.table.getRows as Function) = (reqOpts: any) => {
+      const fn = (reqOpts: any) => {
         assert.deepStrictEqual(reqOpts.filter, expectedFilter);
         assert.strictEqual(FakeMutation.parseColumnName.callCount, 2);
         assert(FakeMutation.parseColumnName.calledWith(keys[0]));
         assert.strictEqual(reqOpts.decode, options.decode);
         done();
       };
+      const row = getRowInstance(fn);
 
       row.get(keys, options, assert.ifError);
     });
@@ -1196,10 +1261,11 @@ describe('Bigtable/Row', () => {
       const expectedFilter = options.filter;
 
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (row.table.getRows as Function) = (reqOpts: any) => {
+      const fn = (reqOpts: any) => {
         assert.deepStrictEqual(reqOpts.filter, expectedFilter);
         done();
       };
+      const row = getRowInstance(fn);
 
       row.get(keys, options, assert.ifError);
     });
@@ -1210,18 +1276,19 @@ describe('Bigtable/Row', () => {
       };
 
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (row.table.getRows as Function) = (reqOpts: any) => {
+      const fn = (reqOpts: any) => {
         assert.strictEqual(reqOpts.decode, options.decode);
         assert(!reqOpts.filter);
         done();
       };
+      const row = getRowInstance(fn);
 
       row.get(options, assert.ifError);
     });
 
     it('should return an error to the callback', done => {
       const error = new Error('err');
-      sandbox.stub(row.table, 'getRows').callsArgWith(1, error);
+      const row = getRowInstanceForErrResp(error as ServiceError);
       row.get((err, row) => {
         assert.strictEqual(error, err);
         assert.strictEqual(row, undefined);
@@ -1230,7 +1297,7 @@ describe('Bigtable/Row', () => {
     });
 
     it('should return a custom error if the row is not found', done => {
-      sandbox.stub(row.table, 'getRows').callsArgWith(1, null, []);
+      const row = getRowInstanceForErrResp(null, []);
       row.get((err, row_) => {
         assert(err instanceof RowError);
         assert.strictEqual(err!.message, 'Unknown row: ' + row.id + '.');
@@ -1245,7 +1312,7 @@ describe('Bigtable/Row', () => {
         a: 'a',
         b: 'b',
       };
-      sandbox.stub(row.table, 'getRows').callsArgWith(1, null, [fakeRow]);
+      const row = getRowInstanceForErrResp(null, [fakeRow]);
       row.get((err, row_) => {
         assert.ifError(err);
         assert.strictEqual(row_, row);
@@ -1263,11 +1330,11 @@ describe('Bigtable/Row', () => {
       };
 
       const keys = ['a', 'b'];
+      const row = getRowInstanceForErrResp(null, [fakeRow]);
 
       row.data = {
         c: 'c',
       };
-      sandbox.stub(row.table, 'getRows').callsArgWith(1, null, [fakeRow]);
       row.get(keys, (err, data) => {
         assert.ifError(err);
         assert.deepStrictEqual(Object.keys(data), keys);